AI Use Case Generator/funnel-ai/backend/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.1
async-timeout==4.0.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.109.2
fonttools==4.60.2
fpdf2==2.8.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
idna==3.11
iniconfig==2.1.0
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.82
langchain-text-splitters==0.3.11
langsmith==0.4.37
Mako==1.3.10
MarkupSafe==3.0.3
nodeenv==1.10.0
openai==1.12.0
orjson==3.11.5
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.9
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.6.1
pytest==8.4.2
python-dotenv==1.0.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
tenacity==9.1.2
tomli==2.3.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.27.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
zstandard==0.25.0
python-multipart
beautifulsoup4
passlib[bcrypt]
python-jose[cryptography]
